backendfiles/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
azure-search-documents==11.4.0
openai>=1.30.0
python-dotenv==1.0.0
azure-identity==1.15.0
pydantic==2.5.0
azure-storage-blob==12.19.0
python-multipart==0.0.9